refactor(wallet): drop switchWalletType and wrap signing in a hook
Remove `switchWalletType` from the useWallet() contract entirely instead
of keeping a no-op stub — a silent no-op is worse than a missing field
because consumers calling it get nothing back with no compile-time
signal. ConnectManagedWalletButton drops the dead "Switch to USD Payments"
branch and renders nothing when the user already has a managed wallet;
ManagedWalletPopup drops its "Switch to Wallet Payments" button (it was
already gated by the self_custody flag that L-4 removes).

Move the signing flow into useSignAndBroadcast — the hook owns the
loading state, snackbar helpers, and TransactionSnackbarContent /
AddCreditsSnackbarContent — while keeping the pure signAndBroadcast
function as the testable core. WalletProvider's body shrinks by ~80
lines.
