+++ apps/deploy-web/src/components/wallet/ConnectManagedWalletButton.tsx
@@ -25,21 +25,21 @@ interface Props extends ButtonProps {
 
 export const ConnectManagedWalletButton: React.FunctionComponent<Props> = ({ className = "", dependencies: d = DEPENDENCIES, ...rest }) => {
   const { settings } = d.useSettings();
-  const { hasManagedWallet, isWalletLoading, switchWalletType } = useWallet();
+  const { hasManagedWallet, isWalletLoading } = useWallet();
   const router = d.useRouter();
 
   const handleClick: React.MouseEventHandler = useCallback(() => {
-    if (hasManagedWallet) {
-      switchWalletType();
-    } else {
-      router.push(UrlService.onboarding());
-    }
-  }, [hasManagedWallet, switchWalletType, router]);
+    router.push(UrlService.onboarding());
+  }, [router]);
+
+  if (hasManagedWallet) {
+    return null;
+  }
 
   return (
     <Button variant="default" onClick={handleClick} className={className} {...rest} disabled={settings.isBlockchainDown || isWalletLoading}>
       {isWalletLoading ? <Spinner size="small" className="mr-2" variant="dark" /> : <Rocket className="rotate-45 text-xs" />}
-      <span className="m-2 whitespace-nowrap">{hasManagedWallet ? "Switch to USD Payments" : "Start Trial"}</span>
+      <span className="m-2 whitespace-nowrap">Start Trial</span>
     </Button>
   );
 };
