+++ apps/deploy-web/src/components/wallet/ManagedWalletPopup/ManagedWalletPopup.tsx
@@ -1,12 +1,11 @@
 import React from "react";
 import { FormattedNumber } from "react-intl";
-import { Button, buttonVariants, Card, CardContent, Separator } from "@akashnetwork/ui/components";
+import { buttonVariants, Card, CardContent, Separator } from "@akashnetwork/ui/components";
 import { cn } from "@akashnetwork/ui/utils";
-import { CoinsSwap, HandCard } from "iconoir-react";
+import { HandCard } from "iconoir-react";
 
 import { useServices } from "@src/context/ServicesProvider";
 import { useWallet } from "@src/context/WalletProvider";
-import { useIsSelfCustodyEnabled } from "@src/hooks/useIsSelfCustodyEnabled";
 import { useManagedEscrowFaqModal } from "@src/hooks/useManagedEscrowFaqModal";
 import type { WalletBalance } from "@src/hooks/useWalletBalance";
 import { LinkTo } from "../../shared/LinkTo";
@@ -21,27 +20,23 @@ export const DEPENDENCIES = {
   Card,
   CardContent,
   Separator,
-  Button,
   LinkTo,
   AddFundsLink,
-  CoinsSwap,
   HandCard,
   FormattedNumber,
   useWallet,
   useManagedEscrowFaqModal,
-  useServices,
-  useIsSelfCustodyEnabled
+  useServices
 };
 
 export const ManagedWalletPopup: React.FC<ManagedWalletPopupProps> = ({ walletBalance, dependencies: d = DEPENDENCIES }) => {
-  const { isManaged, isTrialing, switchWalletType } = d.useWallet();
+  const { isTrialing } = d.useWallet();
   const { showManagedEscrowFaqModal } = d.useManagedEscrowFaqModal();
   const { urlService } = d.useServices();
-  const isSelfCustodyEnabled = d.useIsSelfCustodyEnabled();
 
   return (
     <div className="w-[300px] p-2">
-      {isManaged && isTrialing && (
+      {isTrialing && (
         <div className="mb-2 text-sm font-bold">
           <p className="text-center">Free Trial</p>
         </div>
@@ -85,7 +80,7 @@ export const ManagedWalletPopup: React.FC<ManagedWalletPopupProps> = ({ walletBa
         </d.LinkTo>
       </div>
 
-      {isManaged && isTrialing && (
+      {isTrialing && (
         <div className="my-2 text-center text-xs text-muted-foreground">
           Once your Free credits run out, deployments will automatically close. To continue, create an account and add funds with your credit card. Deployments
           from your Free Trial get transferred when creating a new account.
@@ -100,15 +95,6 @@ export const ManagedWalletPopup: React.FC<ManagedWalletPopupProps> = ({ walletBa
           <HandCard className="text-xs" />
           <span className="whitespace-nowrap">Add Funds</span>
         </d.AddFundsLink>
-        {isSelfCustodyEnabled && (
-          <>
-            <d.Separator className="my-2 bg-secondary/90 dark:bg-white/10" />
-            <d.Button onClick={switchWalletType} variant="outline" className="w-full space-x-2">
-              <d.CoinsSwap />
-              <span>Switch to Wallet Payments</span>
-            </d.Button>
-          </>
-        )}
       </div>
     </div>
   );
